+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
   - https://github.com/Shopify/flash-list/pull/378
 - Change `overrideItemType` prop name to `getItemType`
   - https://github.com/Shopify/flash-list/pull/369
+- Added `useBlankAreaTracker` hook for tracking blank area in production
+  - https://github.com/Shopify/flash-list/pull/411
 - Added `CellRendererComponent` prop
   - https://github.com/Shopify/flash-list/pull/362
 - Added automatic height measurement for horizontal lists even when parent isn't deterministic